requirements.txt
pydantic
pyyaml
rpi5-ws2812
board
gpiozero
adafruit_blinka
dafruit-circuitpython-ads1x15
dafruit-circuitpython-veml7700
